抽离了SubjectSearcher到appdata中
