+++ app/shared/app-data/src/commonMain/kotlin/tools/search/SubjectSearcher.kt
@@ -1,0 +1,73 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.app.tools.search
+
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.map
+import me.him188.ani.app.data.models.ApiResponse
+import me.him188.ani.app.data.models.subject.RatingInfo
+import me.him188.ani.app.data.models.subject.SubjectInfo
+import me.him188.ani.app.data.models.subject.Tag
+import me.him188.ani.app.tools.caching.LazyDataCache
+import me.him188.ani.app.ui.foundation.BackgroundScope
+import me.him188.ani.app.ui.foundation.HasBackgroundScope
+import me.him188.ani.datasources.api.paging.map
+import me.him188.ani.datasources.api.subject.Subject
+import me.him188.ani.datasources.api.subject.SubjectProvider
+import me.him188.ani.datasources.api.subject.SubjectSearchQuery
+import kotlin.coroutines.CoroutineContext
+
+class SubjectSearcher(
+    private val subjectProvider: SubjectProvider,
+    parentCoroutineContext: CoroutineContext,
+) : HasBackgroundScope by BackgroundScope(parentCoroutineContext) {
+    private val currentQuery: MutableStateFlow<SubjectSearchQuery?> = MutableStateFlow(null)
+
+    private val ldc = currentQuery.map { query ->
+        query ?: return@map null
+        LazyDataCache(
+            createSource = { ApiResponse.success(subjectProvider.startSearch(query).map { it.toSubjectInfo() }) },
+            getKey = { it.id },
+            debugName = "SubjectSearcher.ldc",
+        )
+    }.shareInBackground(started = SharingStarted.Lazily)
+
+    val list = ldc.flatMapLatest { it?.cachedDataFlow ?: flowOf(emptyList()) }
+    val hasMore = ldc.flatMapLatest { it?.isCompleted ?: flowOf(true) }
+        .map { !it }
+
+    suspend fun requestMore() = ldc.first()?.requestMore()
+
+    fun clear() {
+        currentQuery.value = null
+    }
+
+    fun search(query: SubjectSearchQuery) {
+        currentQuery.value = query
+    }
+}
+
+
+fun Subject.toSubjectInfo(): SubjectInfo {
+    return SubjectInfo(
+        id = id,
+        name = originalName,
+        nameCn = chineseName,
+        summary = this.summary,
+        tags = this.tags.map { Tag(it.first, it.second) }.sortedByDescending { it.count },
+        imageCommon = this.images.landscapeCommon,
+        imageLarge = this.images.largePoster,
+        ratingInfo = RatingInfo.Empty.copy(rank = rank),
+    )
+}
