+++ app/shared/src/commonMain/kotlin/ui/home/search/SearchViewModel.kt
@@ -31,64 +31,22 @@ import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import me.him188.ani.app.data.models.ApiResponse
 import me.him188.ani.app.data.models.preference.OneshotActionConfig
 import me.him188.ani.app.data.models.preference.SearchSettings
-import me.him188.ani.app.data.models.subject.RatingInfo
-import me.him188.ani.app.data.models.subject.SubjectInfo
-import me.him188.ani.app.data.models.subject.Tag
 import me.him188.ani.app.data.persistent.database.eneity.SearchHistoryEntity
 import me.him188.ani.app.data.persistent.database.eneity.SearchTagEntity
 import me.him188.ani.app.data.repository.SettingsRepository
 import me.him188.ani.app.data.repository.SubjectSearchRepository
-import me.him188.ani.app.tools.caching.LazyDataCache
+import me.him188.ani.app.tools.search.SubjectSearcher
 import me.him188.ani.app.ui.foundation.AbstractViewModel
-import me.him188.ani.app.ui.foundation.BackgroundScope
-import me.him188.ani.app.ui.foundation.HasBackgroundScope
-import me.him188.ani.datasources.api.paging.map
-import me.him188.ani.datasources.api.subject.Subject
 import me.him188.ani.datasources.api.subject.SubjectProvider
 import me.him188.ani.datasources.api.subject.SubjectSearchQuery
 import me.him188.ani.utils.coroutines.update
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import kotlin.coroutines.CoroutineContext
-
-class SubjectSearcher(
-    private val subjectProvider: SubjectProvider,
-    parentCoroutineContext: CoroutineContext,
-) : HasBackgroundScope by BackgroundScope(parentCoroutineContext) {
-    private val currentQuery: MutableStateFlow<SubjectSearchQuery?> = MutableStateFlow(null)
-
-    private val ldc = currentQuery.map { query ->
-        query ?: return@map null
-        LazyDataCache(
-            createSource = { ApiResponse.success(subjectProvider.startSearch(query).map { it.toSubjectInfo() }) },
-            getKey = { it.id },
-            debugName = "SubjectSearcher.ldc",
-        )
-    }.shareInBackground(started = SharingStarted.Lazily)
-
-    val list = ldc.flatMapLatest { it?.cachedDataFlow ?: flowOf(emptyList()) }
-    val hasMore = ldc.flatMapLatest { it?.isCompleted ?: flowOf(true) }
-        .map { !it }
-
-    suspend fun requestMore() = ldc.first()?.requestMore()
-
-    fun clear() {
-        currentQuery.value = null
-    }
-
-    fun search(query: SubjectSearchQuery) {
-        currentQuery.value = query
-    }
-}
 
 @Stable
 class SearchViewModel(
@@ -206,19 +164,6 @@ class SearchViewModel(
             ),
         )
     }
-}
-
-fun Subject.toSubjectInfo(): SubjectInfo {
-    return SubjectInfo(
-        id = id,
-        name = originalName,
-        nameCn = chineseName,
-        summary = this.summary,
-        tags = this.tags.map { Tag(it.first, it.second) }.sortedByDescending { it.count },
-        imageCommon = this.images.landscapeCommon,
-        imageLarge = this.images.largePoster,
-        ratingInfo = RatingInfo.Empty.copy(rank = rank),
-    )
 }
